rollbak migration transactions hook and refactor
